fix(schema): local_order_id INTEGER -> BIGINT em robot_signals_log + intent
Bug descoberto durante validacao do P0 #2 fix em ambiente real (smoke
05/mai 10:30): agent retorna local_order_id no formato yymmdd+ms+counter
(ex: 126050507283019, ~1.3x10^14). INT4 max e' 2.1x10^9 — overflow garante
que TODO update_intent_sent/update_signal_log_sent falha com 'integer out
of range', alimentando o warning intent_update_skipped que fechei em P0 #2.

Por que nao quebrou em smoke 04/mai? Naquela sessao apenas 25 ordens manuais
foram disparadas e nem todas chegaram a tentar update via dispatcher. Smoke
05/mai exercitou auto_trader em loop (1 dispatch/min) — overflow recorrente.

Aplicado ALTER em DB market_data:5433 (Timescale, onde robot_* moram de fato
— nao confundir com tabelas zumbi em postgres:5432 conforme aviso ts_0004):
  ALTER TABLE robot_orders_intent ALTER COLUMN local_order_id TYPE BIGINT;
  ALTER TABLE robot_signals_log   ALTER COLUMN local_order_id TYPE BIGINT;

Validado em prod: dispatch 10:55:47 ok sem nenhum update_*_failed.

init_timescale/006 atualizado pra futuros containers nascerem corretos.
Sem migration alembic ts_0005 — DDL de ramo timescale fica em
init_timescale por convencao (vide aviso em ts_0004).

+++ init_timescale/006_robot_trade.sql
@@ -44,7 +44,9 @@ CREATE TABLE IF NOT EXISTS robot_signals_log (
     computed_at         TIMESTAMPTZ     NOT NULL DEFAULT NOW(),
     -- Quando enviado: liga ao callback DLL via local_order_id
     sent_to_dll         BOOLEAN         NOT NULL DEFAULT FALSE,
-    local_order_id      INTEGER,         -- preenche apos POST /agent/order/send retornar
+    -- BIGINT: agent retorna IDs ate ~10^14 (formato yymmdd+ms+counter); INT4
+    -- overflowa em local_order_id > 2.1B (smoke 05/mai 10:37 quebrou aqui).
+    local_order_id      BIGINT,         -- preenche apos POST /agent/order/send retornar
     -- Quando NAO enviado: motivo (kill switch, risk reject, no signal, etc.)
     reason_skipped      TEXT,
     -- Snapshot do contexto p/ debug (preco, sinal_ml, regimen, position size)
@@ -77,8 +79,9 @@ CREATE TABLE IF NOT EXISTS robot_orders_intent (
     -- OCO attached (TP + SL): NULL se ordem solta
     take_profit         DOUBLE PRECISION,
     stop_loss           DOUBLE PRECISION,
-    -- Liga ao callback DLL (preenchido apos resposta /order/send)
-    local_order_id      INTEGER,
+    -- Liga ao callback DLL (preenchido apos resposta /order/send).
+    -- BIGINT: ver nota em robot_signals_log.local_order_id acima.
+    local_order_id      BIGINT,
     cl_ord_id           TEXT,            -- idempotencia no proxy
     -- Status interno (sucesso/erro envio, NAO o status DLL final)
     sent_at             TIMESTAMPTZ,
